fix: Completely update custom.css and mobile.css files
This commit completely removes all code from custom.css
and mobile.css. Why? The previous code had custom css from
the Dawn theme that interfered with other themes.

So we had to start from scratch and update only the css
needed for each liquid section, not the whole theme.

Which is what I did.

THE UPDATED CSS ONLY WORKS FOR 'GALLERY-TABS.LIQUID' AND
'FLICKITY-UGC-SLIDER.LIQUID'. That's it. CSS updates
for the rest of the liquid files coming soon.

Finally, I also updated the README file to reflect these
new changes.